requirements.txt
psycopg[binary]==3.2.3
requests==2.32.3
python-dateutil==2.9.0.post0